requirements.txt
flask-apispec
marshmallow-sqlalchemy
Flask-SQLAlchemy
flask-marshmallow
pytest
Authlib